pyproject.toml
[project]
name = "searxng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp-searxng>=0.1.0",
]
